refactor layout inflalter for Recycler view adapters (code climate issue)
