+++ app/src/main/java/com/github/sdpteam15/polyevents/adapter/EventItemAdapter.kt
@@ -1,7 +1,5 @@
 package com.github.sdpteam15.polyevents.adapter
 
-import android.os.Build
-import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
@@ -47,9 +45,9 @@ class EventItemAdapter(
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ItemViewHolder {
-        val adapterLayout = LayoutInflater.from(parent.context)
-            .inflate(R.layout.tab_event, parent, false)
-        return ItemViewHolder(adapterLayout)
+        return ItemViewHolder(HelperAdapterFunctions.createAdapterLayout(
+            parent, R.layout.tab_event
+        ))
     }
 
     override fun onBindViewHolder(holder: ItemViewHolder, position: Int) {
