+++ app/src/main/java/com/github/sdpteam15/polyevents/adapter/ItemAdapter.kt
@@ -1,6 +1,5 @@
-package com.github.sdpteam15.polyevents.item_request
+package com.github.sdpteam15.polyevents.adapter
 
-import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageButton
@@ -53,9 +52,9 @@ class ItemAdapter (
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ItemViewHolder {
-        val adapterLayout = LayoutInflater.from(parent.context)
-            .inflate(R.layout.tab_material_item, parent, false)
-        return ItemViewHolder(adapterLayout)
+        return ItemViewHolder(HelperAdapterFunctions.createAdapterLayout(
+            parent, R.layout.tab_material_item
+        ))
     }
 
     override fun onBindViewHolder(holder: ItemViewHolder, position: Int) {
